CDRIVER-5859 Integer comparison macro (#1845)
* Integer type and range utilities.

This will lay the basis for more convenient checked
arithmetic going forward.

* mlib/config.h header for common platform detection and utility macros
* Integer comparison utilities.

These are intended to replace existing integer comparison functions/macros.
Having different comparison functions for differing sign/size combinations is
error-prone, resists refactoring, and requires programmer diligence at call
sites. `mlib_cmp` is a single macro that works with all integer types,
regardless of sign or size.

`mlib_cmp` internally branches on the signedness properties of the operands. In
unoptimized cases, this may be slower, but any inliner should trivially
eliminate the dead code branches.

* Mass update: Replace `mcommon_cmp` with `mlib_cmp`
* Rename variable to not fight with clang-format
* Mark mlib_cmp as force-inline
* Add a link-time check in release builds

If we are building in optimized mode, we check that any calls to
`mlib_cmp` are inlined and that DCE passes over the inlined result.

* Move mcommon tests into mlib test file