Fix BSON_NORETURN for C23 standard
The new BSON_NORETURN macro has implementations for C11, C23, GCC, and a fallback.
The C23 version was untested in CI, and led to CDRIVER-5889 when the about-to-be-released
GCC 15 switched to the C23 standard as default.

Failures are also observed in normal libmongoc build with a new enough gcc and -DCMAKE_C_STANDARD=23

With this patch, all branches of the BSON_NORETURN definition are working.