test-bson-vector read/write edge cases use malloc'ed value buffers
The main reason for this change is to avoid an unhelpful implementation of -Warray-bounds in gcc 11
which notices that the memcpy in these tests can be out-of-range, but doesn't notice that the out-of-range
memcpy will never be executed. The easiest fix was to dynamically allocate the value buffers, to prevent
gcc from associating range info with these pointers.

This replaces some arbitrary nonzero test values with zeroes, out of convenience.