Don't update VERSION_CURRENT during fetch-source
fetch-source would pull sources from the repo, and
then recalculate the current version and overwrite
the VERSION_CURRENT file. We will rely on VERSION_CURRENT
going forward.

fetch-source also tried ot create a qualified name for the commit as
$CURRENT_VERSION, but this was bugged and only included the commit hash on patch
builds. We don't rely on this, so we'll go with just using the commit hash for
now.