[CDRIVER-5961] Renaming of binaries, packages, and install paths (#1965)
Resolves CDRIVER-5961 and CDRIVER-2086

* Add CTest cases to test pkg-config importing

These use the TestProject module to define test
cases that run a CMake build, but that CMake project
uses `pkg-config` to import the libraries rather than `find_package`.
This allows us to test `pkg-config` while relying on CMake to do the
heavy lifting around compiler detection and search path handling.

* Set CMAKE_PREFIX_PATH in TestProject

When building a TestProject, we want to set CMAKE_PREFIX_PATH to point
to the directory that received the parent project installation.

* Change the link line for pkg-config static libs

Previously, pkg-config generated a `-l` with a library stem, and it
would only coincidentally name a static library with a `.a` suffix.
Instead, for static libraries, specify the path to the actual static
library archive file. Only rely on `-l` searching for shared libraries.

* Rename pkg-config files

* Update file names and install paths

This change is BREAKING and changes the names of generated files for
libbson and libmongoc.

  * Headers are installed in a directory qualified by the project version.
  * The generated dynamic library has a property version suffix now (instead of 0.0.0).
  * We set the SOVERSION property to generate a new SONAME for the emitted bson library.
  * The bson-1.0 and mongoc-1.0 CMake packages are removed.
  * The pkg-config files are now named `bson1` / `bson1-static` and `mongoc1` / `mongoc1-static`.
  * Both the static and dynamic library file stems are now `libbson1` / `libmongoc1` (no
  `-static` suffix).
  * The project now imports the system libbson using the name "bson" instead
  of `bson-1.0`

* Remove fragile installed-files tests, revise some for new filenames

* Give DLL implibs a different file extension on Windows

* Remove install code for Apple frameworks and modulemaps

* Remove libmongoc-ssl .pc files

* Add NEWS entries regarding package renames

* Qualify installation of docs and licenses

* Major-version qualify the `mongoc-stat` tool