CDRIVER-5956 raise min wire version from 7 to 8 (server 4.2) (#1998)
* bump WIRE_VERSION_MIN to 4.2

* update NEWS

* remove test tasks for 4.0

* regenerate Evergreen config

* remove `WIRE_VERSION_4_0`

Remove behavior and tests that only run on server <= 4.0.

* remove transactions wireversion=7 check

* remove test check for wireversion=5

* remove change stream prose test 9

Test is only expected to run on server >=4.0 and <4.0.7.

* expect `killCursors` in resume test

The "not primary" error does not cause the connection to be closed in > 4.0. Expect a `killCursors` command is sent on resume.

* remove prose test 12

Test requires server < 4.0.7

* remove unnecessary test wire version checks

Tests are expected to run on all supported server versions.

* remove test checks for wireversion=7

* comment describing server version for wire version

* remove test checks for wireversion<8

* remove test check for wireversion=6

* remove test checks for server 3.8

* skip pre-4.2 SDAM tests

* skip auth tests using 4.0 server

* update exhaust cursor tests to expect OP_MSG

Expect `killCursors` command on non-network error to avoid delay in `mongoc_cursor_destroy`.