A parallel formatting and include-fixup script
This adds a `format.py` script that formats files
with `clang-format`, as well as fixing up `#include`
directives to use `<angle>` brackets if they match
a certain pattern.

This script also runs all jobs in parallel, making it
significantly faster that a single big `clang-format`
command, and more suitable as a quick precommit check.