Prepare some core header files for public use.
This change starts a new pattern for public API
headers: Headers that omit the "bson-" or "mongoc-"
prefix are intended for public `#include` directives, allowing users
to selectively import the APIs that they want.

This change required minimal modification of the existing code, as the
few headers that have been modified are very basic and have very few
dependencies.