Add CMake code for header verification.
This adds a CMake utility based around
VERIFY_INTERFACE_HEADER_SETS to check that certain headers can be
included directly in a translation unit without any prerequesits, in
either C or C++ code.

This verification only concerns itself with headers that we have already
made part of the public API. Private header files and header files
that are not-yet-sanitized are not checked by this process.

The verification does not yet happen automatically. It is only enabled
for CMake 3.24 or newer, and requires the dev/CI to build the special
linting target the CMake generates.