CDRIVER-5998 share `SCHANNEL_CRED` (#2052)
* add `const` to `mongoc_ssl_opt_t` params

* rename `mongoc_secure_channel_cred` to `mongoc_secure_channel_cred_handle`

* share `SCHANNEL_CRED`

Share `SCHANNEL_CRED` on all client/pool connections. This is intended to simplify future fix for CDRIVER-5998. As an added benefit, this reduces the repeated reading of PEM files for each connection.

* fix X509 test

Capture logs earlier to account for secure channel loading the PEM file when SSL options are set. This revealed the OpenSSL implementation also logs.

* fix existing memory leak on failure