[CDRIVER-5670] Include-directive Sorting and Formatting (#2050)
* Set a stable header sorting, and fixup errors for it

This sets a new header ordering that is maintained
by `clang-format`. They are sorted based on logical
code locallity, to reduce further implicit dependencies in the future.

Some code needed to be fixed after header sorting, since it relied on
implicit ordering. Those changes are relatively minor.

* A parallel formatting and include-fixup script

This adds a `format.py` script that formats files
with `clang-format`, as well as fixing up `#include`
directives to use `<angle>` brackets if they match
a certain pattern.

This script also runs all jobs in parallel, making it
significantly faster that a single big `clang-format`
command, and more suitable as a quick precommit check.

The #include directives are now rewritten according to
the rules:

1. If the header uses quotes and the first filename char is a dot, it is
   retained.
2. If the quoted header path specifies a file relative to the file in
   which it appears, the quoted style is kept, and a `./` is prepended
   to the path.
3. Otherwise, the directive is rewritten to use `<angles>`

`clang-format` now sorted quote-includes to the top of the file, assuming
they are maximally local for the including file.

* Use new formatting script for formatting tasks
* No conditional stdbool.h: We require C99