Remove the `Language: Cpp` setting for formatting (#2055)
* Remove the `Language: Cpp` setting for formatting

It seems that the `Language: Cpp` setting in `.clang-format` changes Clang's behavior
in Clang 20.x, and sometimes restricts it from formatting C files, instead falling back
to the default format.

This change removes `Language: Cpp` and allows `clang-format` to apply the same
formatting to all source files. This seems to also have minor changes to the
way it typesets the `not` identifier in C files (it looks like `clang-format` was
previously treating C as C++, where `not` is a special prefix operator rather 
than a bare identifier).

---------

Co-authored-by: Ezra Chung <88335979+eramongodb@users.noreply.github.com>