CDRIVER-6053 do not buffer pooled streams (#2064)
* disable buffering for streams

The stream buffer is tied to the lifetime of a stream. A large response may grow the buffer and result in higher memory usage until the stream is destroyed. A stream may live as long as the client/pool.

The buffer appears unnecessary, since a separate buffer is allocated, and read directly into, for the server responses.

* remove unnecessary `buffered` parameter from `mongoc_client_connect`

* warn about allocation strategy of buffered stream

Co-authored-by: Ezra Chung <88335979+eramongodb@users.noreply.github.com>