[CDRIVER-4153] Beginning of Header Hygiene & Verification (#2053)
* Prepare some core header files for public use.

This change starts a new pattern for public API
headers: Headers that omit the "bson-" or "mongoc-"
prefix are intended for public `#include` directives, allowing users
to selectively import the APIs that they want.

This change required minimal modification of the existing code, as the
few headers that have been modified are very basic and have very few
dependencies.

* Fix C++ and C compatibility in some header files
* Add CMake code for header verification.

This adds a CMake utility based around
VERIFY_INTERFACE_HEADER_SETS to check that certain headers can be
included directly in a translation unit without any prerequesits, in
either C or C++ code.

This verification only concerns itself with headers that we have already
made part of the public API. Private header files and header files
that are not-yet-sanitized are not checked by this process.

The verification does not yet happen automatically. It is only enabled
for CMake 3.24 or newer, and requires the dev/CI to build the special
linting target the CMake generates.

* Earthly target and EVG tasks to run header verification
* Remove unused old script
* Rename bson.h -> bson-bcon.h