Fix asserting arithmetic on big-endian platforms
Broken: If the target type of a ckdint-asserting-arithmetic
macro is smaller than intmax_t, then the return value of the function
will be bogus on big-endian platforms, because we neglected to adjust
the temporary bits used to store the result.

Fix: If the platform is big-endian, we bit-shift the temporary result
down by the number of bytes of different between uintmax_t and the
destination type size.