[CDRIVER-6142] Refactor dependency/environment setup in Earthfile (#2165)
* Consolidate several small scripts into a single __tool
* Significant Earthfile refactor

This change rewrites much of Earthfile's dependency-installation handling
code to be simpler to modify, and more flexible to use.

Previously, each environment had its own target that needed to know how
to install all of our dependencies for the specific platform. This
refactor inverts the relationship: Instead, each dependency has an Earthly
function that knows how to install itself based on whatever platform it
happens to be running within.

This allows the FROM line to be a plain image, rather that one of our
handcrafted `env.foo` targets. This will make introducing and testing
additional environments far more simple in the future.

This change also allows us to remove the tools/build.earth utility, since we can
now fit everything in the top-level file.

* Update Earthly-generated tasks
* Add Ubuntu 24.04
* Enable building both with/without Snappy in Earthly
* Tasks with/without Snappy
* Add a trailhead for obtaining the detailed Earthfile documentation
* Refactor Earthly C++ installation to rely on the same C compiler param