CDRIVER-6153 run OIDC unified test with GCP (#2188)
* support "mongodb+srv" in `call_hello`

* do not log test URI or credentials

Evergreen logs are private, but this is intended to reduce the scope of credentials if `test-libmongoc` is run with a sensitive URI.

* do not log Unix domain socket path

Causes a failure if `MONGODB_TEST_URI` is set to a "mongodb+srv://" URI due to a call to `test_framework_get_port`. Logging the domain socket URI is not expected to be very useful.

* handle "mongodb+srv" URI in `test_framework_add_user_password`

* support GCP in unified tests

document `MONGOC_GCP_RESOURCE` environment variable

* remove more logs of URIs

* suggest not logging in CONTRIBUTING.md

* fix uninitialized free on error in example

* consistently use `fprintf` in example errors